requirements.txt
requests==2.31.0
pandas==2.1.4
nltk==3.8.1
gensim==4.3.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
python-dotenv==1.0.1
scipy==1.11.4
